Account for timezone name changes when building timezone lookup
This fixes a bug where countries that have deprecated timezone names
end up with no timezones at all!

As an example, Cambodia's old timezone used to be 'Asia/Phnom_Pehn',
but that was deprecated in favor of 'Asia/Bangkok'.

`TZPicker.pm` gets its timezone names from the array returned by
`DateTime::TimeZone->all_names`, but that array does not include old
timezone names.

In order to make the `_build_zone_lookup` subroutine account for
timezone names that have changed, we need to update this array with the
new timezone names where appropriate and use that instead. Luckily,
`DateTime::TimeZone->links` provides us with a hash that maps all of the
old timezone names to the new ones.